+++ src/components/UserDetailSections/UserInfo/hooks/index.js
@@ -1,0 +1,1 @@
+export { default as useProfilePicture } from './useProfilePicture';